requirements.txt
# =============================================================================
# DuoNova Resume Processing Application Requirements
# Updated: August 2025
# =============================================================================

# -----------------------------------------------------------------------------
# Core Web Framework
# -----------------------------------------------------------------------------
fastapi==0.116.1               # Modern, fast web framework for building APIs
uvicorn[standard]==0.35.0      # ASGI server implementation
starlette==0.47.3              # Lightweight ASGI framework/toolkit

# -----------------------------------------------------------------------------
# Database & Storage
# -----------------------------------------------------------------------------
motor==3.7.1                   # Async MongoDB driver
pymongo==4.14.1               # MongoDB driver for Python
dnspython==2.7.0               # Required for MongoDB Atlas connections

# -----------------------------------------------------------------------------
# AWS Cloud Services
# -----------------------------------------------------------------------------
boto3==1.40.18                 # AWS SDK for Python
botocore==1.40.18             # Low-level interface to AWS services
s3transfer==0.13.1            # S3 transfer utilities

# -----------------------------------------------------------------------------
# File Processing & Document Handling
# -----------------------------------------------------------------------------
PyPDF2==3.0.1                 # PDF reading and manipulation
python-docx==1.2.0            # Microsoft Word document processing
lxml==6.0.1                   # XML and HTML processing library
python-multipart==0.0.20      # File upload support for FastAPI

# -----------------------------------------------------------------------------
# LLM & AI Providers
# -----------------------------------------------------------------------------
openai==1.102.0               # OpenAI API client
groq==0.31.0                  # Groq API client for fast inference

# -----------------------------------------------------------------------------
# Data Validation & Configuration
# -----------------------------------------------------------------------------
pydantic==2.11.7              # Data validation using Python type annotations
pydantic_core==2.33.2         # Core validation logic for Pydantic (compatible with pydantic 2.11.7)
python-dotenv==1.1.1          # Load environment variables from .env file
email-validator==2.3.0        # Email validation for Pydantic

# -----------------------------------------------------------------------------
# HTTP & Networking
# -----------------------------------------------------------------------------
httpx==0.28.1                 # Async HTTP client library
httpcore==1.0.9               # Core HTTP functionality
requests==2.32.5              # HTTP library for Python

# -----------------------------------------------------------------------------
# Text Processing & Utilities
# -----------------------------------------------------------------------------
regex==2025.7.34              # Alternative regular expression module

# -----------------------------------------------------------------------------
# Development & Testing Dependencies
# -----------------------------------------------------------------------------
pytest==7.4.3                 # Testing framework
pytest-asyncio==0.21.1        # Async support for pytest

# -----------------------------------------------------------------------------
# Core System Dependencies
# -----------------------------------------------------------------------------
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
distro==1.9.0
h11==0.16.0
httptools==0.6.4
idna==3.10
jiter==0.10.0
jmespath==1.0.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0
watchfiles==1.1.0
websockets==15.0.1
